+++ src/video/video.service.ts
@@ -1,4 +1,8 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
+import {
+	ForbiddenException,
+	Injectable,
+	NotFoundException,
+} from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Video } from "./entities/video.entity";
 import { Repository } from "typeorm";
@@ -34,14 +38,16 @@ export class VideoService {
 		return video;
 	}
 
-	async update(id: number, updateVideoDto: UpdateVideoDto) {
-		await this.findOne(id);
+	async update(id: number, authorId: number, updateVideoDto: UpdateVideoDto) {
+		const video = await this.findOne(id);
+		if (video.author.id !== authorId) throw new ForbiddenException();
 		await this.videoRepository.update(id, updateVideoDto);
 		return this.findOne(id);
 	}
 
-	async remove(id: number) {
+	async remove(id: number, authorId: number) {
 		const video = await this.findOne(id);
+		if (video.author.id !== authorId) throw new ForbiddenException();
 		await this.videoRepository.remove(video);
 		return { deleted: true };
 	}
